refactor: extract evidence ingress routes (#1733)
